% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worm.R
\docType{data}
\name{StatWorm}
\alias{StatWorm}
\alias{stat_worm}
\alias{geom_worm}
\title{Worms}
\format{
An object of class \code{StatWorm} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatWorm

stat_worm(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  shorten_lines = 1,
  n = 100,
  dodge = 0,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_worm(
  mapping = NULL,
  data = NULL,
  stat = "worm",
  position = "identity",
  shorten_lines = 1,
  n = 100,
  dodge = 0,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{shorten_lines}{How much space should be given for smooth transitions between
steps? \code{0} results in a regular stepped line. Default is 1. Smoothing is capped at
half the width of the shortest adjacent step.}

\item{n}{The number of points drawn for each Bezier curve. In other words: How smooth
or jagged are the transitions? Default is 100.}

\item{dodge}{Worm steps less than this distance apart will be grouped and dodged on
the y axis. 0 results in no dodging (default). Dodging is implemented here as part of
the stat (rather than as a position function) so that Bezier curves can be drawn afterward.}

\item{arrow}{Arrow specification, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}
}
\description{
The worm geom is a stepped line with smooth, aesthetic transitions between steps.
}
\details{
Data must be formatted with one data point for each new step, plus one extra for the
end of each worm. Worms are distinguished by the group aesthetic.
}
\section{Aesthetics}{

geom_worm understands the following aesthetics (required aesthetics are in bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{group}
\item color
\item linewidth
\item linetype
\item alpha
\item lineend
}
}

\examples{
   # one data point for each new step, plus one extra for the end of the worm
   data <- data.frame(
      x = c(1, 2, 4, 1, 3, 4, 1, 1.5, 2.5, 4),
      y = c(1, 2, 2, 2, 1, 1, 1.5, 1, 1.7, 1.7),
      group = rep(c("you", "me", "him"), times = c(3, 3, 4))
   )

   ggplot(data, aes(x, y, group = group, color = group)) +
    geom_worm(linewidth = 10, shorten_lines = .2, dodge = 0.1)
}
\keyword{datasets}
