% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nobel_physicists}
\alias{nobel_physicists}
\title{Lives of Winners of the Nobel Prize for Physics, 1901-1925}
\format{
A data frame with 94 rows and 4 variables: \describe{
\item{name}{Character Full name of physicist}
\item{country}{Factor. Country of residence at given date}
\item{year}{Integer}
\item{nobel_prize}{Integer. Year awarded to Nobel Prize}
}
}
\usage{
nobel_physicists
}
\description{
A dataset containing information about the residences of 30 prize-winning scientists
over the course of their lifetimes. Data from Wikipedia.
}
\details{
For details on formatting see \link{geom_worm}.
}
\keyword{datasets}
