% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormsplot.R
\name{wormsplot}
\alias{wormsplot}
\title{Automatically Generate a Plot of Historical Figures as They Move From Region to Region}
\usage{
wormsplot(
  data,
  x = "x",
  region = "y",
  id = "id",
  label.worms = TRUE,
  label.args = list(),
  worm.color = "initial",
  worm.args = list(linewidth = 10),
  backplate.args = list(alpha = 0.3),
  region.spacing = 2,
  x.padding = 1,
  region.label.width = 4,
  region.label.args = list(),
  x.breaks.by = 10
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{x}{string. The name of the variable to be plotted on the x axis.}

\item{region}{string. The name of the categorical variable to be plotted on the y axis.}

\item{id}{string. The name of the grouping variable.}

\item{label.worms}{If \code{TRUE} (default) worms will be labeled with their id}

\item{label.args}{list. additional parameters to be passed to \code{geom_text()}}

\item{worm.color}{How should the worms be colored? If 'initial' (default), color is according
to their region of origin. If 'final', color will be according to their final destination.
'id' will give each worm its own color.}

\item{worm.args}{list. additional parameters to be passed to \code{stat_worm()}}

\item{backplate.args}{list. additional parameters to be passed to \code{geom_tile()} for creating
colored boxes underneath the plot.}

\item{region.spacing}{How much padding be added to regions on the y axis?}

\item{x.padding}{Extra space beyond the highest and lowest plotted values on the x axis.}

\item{region.label.width}{The horizontal width of colored boxes under the region labels.}

\item{region.label.args}{list. additional parameters to be passed to \code{geom_tile()} for creating
colored boxes underneath region labels.}

\item{x.breaks.by}{Frequency of breaks on the x axis. Defaults to 10.}
}
\description{
\code{wormsplot()} is a convenience function that automatically dodges worms along the
y axis and scales regions accordingly. It comes with default settings for creating
an aesthetic historical visualization. Further customization can by applied to
the resulting ggplot object.
}
\examples{
  data <- data.frame(
    date = c(0, 30,
             1, 10, 25, 30,
             1, 9, 14, 30),
    place = c('Dagobah', 'Dagobah',
              'Hoth', 'Dagobah', 'Bespin', 'Bespin',
              'Hoth', 'Asteroids', 'Bespin', 'Bespin'),
    person = c('Yoda', 'Yoda',
               'Luke', 'Luke', 'Luke', 'Luke',
               'Princess Leia', 'Princess Leia', 'Princess Leia', 'Princess Leia')
  )

  wormsplot(data, 'date', 'place', 'person')
}
