% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormsplot.R
\name{wormsplot}
\alias{wormsplot}
\title{Automatically Generate a Plot of Historical Figures as They Move From Region to Region}
\usage{
wormsplot(
  data,
  x = "x",
  region = "y",
  id = "id",
  label.worms = TRUE,
  shorten_lines = 1,
  n = 50,
  ...,
  region.spacing = 2,
  color = "initial",
  linewidth = 10,
  x.padding = 1,
  region.label.width = 4,
  x.breaks.by = 10,
  backplate.alpha = 0.3
)
}
\arguments{
\item{data}{a dataframe or tibble}

\item{x}{string. The name of the variable to be plotted on the x axis.}

\item{region}{string. The name of the categorical variable to be plotted on the y axis.}

\item{id}{string. The name of the grouping variable.}

\item{label.worms}{If \code{TRUE} (default) worms will be labeled with their id}

\item{shorten_lines}{How much space should be given for smooth transitions between
steps? \code{0} results in a regular stepped line. Default is 1. Smoothing is capped at
half the width of the shortest adjacent step.}

\item{n}{The number of points drawn for each Bezier curve. In other words: How smooth
or jagged are the transitions? Default is 100.}

\item{...}{additional parameters to be passed to \code{geom_worm()} and \code{geom_text()}}

\item{region.spacing}{How much padding be added to regions on the y axis?}

\item{color}{How should the worms be colored? If 'initial' (default), color is according
to their region of origin. If 'final', color will be according to their final destination.
'id' will give each worm its own color.}

\item{linewidth}{The width of the worms.}

\item{x.padding}{Extra space beyond the highest and lowest plotted values on the x axis.}

\item{region.label.width}{The horizontal width of colored boxes under the region labels.}

\item{x.breaks.by}{Frequency of breaks on the x axis. Defaults to 10.}

\item{backplate.alpha}{Transparency of colored boxes underneath the plot. Defaults to 0.3.}
}
\description{
\code{wormsplot()} is a convenience function that automatically dodges worms along the
y axis and scales regions accordingly. It comes with default settings for creating
an aesthetic historical visualization. Further customization can by applied to
the resulting ggplot object.
}
\examples{
  data <- data.frame(
    date = c(0, 30,
             1, 10, 25, 30,
             1, 9, 14, 30),
    place = c('Dagobah', 'Dagobah',
              'Hoth', 'Dagobah', 'Bespin', 'Bespin',
              'Hoth', 'Asteroids', 'Bespin', 'Bespin'),
    person = c('Yoda', 'Yoda',
               'Luke', 'Luke', 'Luke', 'Luke',
               'Princess Leia', 'Princess Leia', 'Princess Leia', 'Princess Leia')
  )

  wormsplot(data, 'date', 'place', 'person')

  data(nobel_physicists)
  nobel_physicists \%>\%
    filter(name \%in\% head(unique(name), 10)) \%>\%
    mutate(country = forcats::fct_drop(country)) \%>\%
    wormsplot('year', 'country', 'name', color = 'initial',
    linewidth = 5.5, region.label.width = 18, size = 3.6) +
    labs(title = "The Lives of Winners of the Nobel Prize in Physics 1901-1907")
}
